iles-main/requirements.txt
Django==5.2.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.2
Pillow==11.2.1
